Move degrees into shared
